feat: renaming (remove StorageHost~)
